Changed TipAmount and TipPercent type from decimal? to double?

+++ Src/Server/MK.Booking/MK.Booking.Commands/CreateOrder.cs
@@ -52,8 +52,8 @@ namespace apcurium.MK.Booking.Commands
         {
             public bool PayWithCreditCard { get; set; }
             public Guid CreditCardId { get; set; }
-            public decimal? TipAmount { get; set; }
-            public decimal? TipPercent { get; set; }
+            public double? TipAmount { get; set; }
+            public double? TipPercent { get; set; }
         }
  
     }
