+++ Src/Server/MK.Booking/MK.Booking/Domain/PaymentInformation.cs
@@ -5,7 +5,7 @@ namespace apcurium.MK.Booking.Domain
     public class PaymentInformation
     {
         public Guid CreditCardId { get; set; }
-        public decimal? TipAmount { get; set; }
-        public decimal? TipPercent { get; set; }
+        public double? TipAmount { get; set; }
+        public double? TipPercent { get; set; }
     }
 }
